Separate XML lang parser into its own class
